src/requirements.txt
defusedxml==0.6.0
diff-match-patch==20200713
# Django==2.2.14
Django==3.2.16
django-appconf==1.0.5
# django-betterforms==1.2
django-betterforms==2.0.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-crispy-forms==1.13.0
django-imagekit==4.1.0
django-import-export==2.5.0
django-mysql==3.10.0
django-session-timeout==0.1.0
django-webpack-loader==0.7.0
django-widget-tweaks==1.4.3
django-widgets-improved==1.5.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.6
pilkit==2.0
Pillow==9.1.0
python-dateutil==2.8.2
pytz==2020.4
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
ua-parser==0.10.0
user-agents==2.1
xlrd==2.0.1
xlwt==1.3.0
gunicorn==20.1.0